Add vim-plug plugin autoloader for vim
To enable autoloading Vim plugins; my current vimrc will emit errors if
plugins are not downloaded and initialised via this plugin.
Unfortunately it will need to be be updated manually, and a separate
folder vim/plugged/ will need to be created manually for the plugins to
be installed to, but I plan to automate this with dotbot in the near
future.